Extract matcher to separate files
