+++ android/app/src/androidTest/java/com/example/domag/NotificationsUiTest.kt
@@ -1,0 +1,2 @@
+package com.example.domag
+
